Split early gmregclear code into its own TU
Keep the stage tables with the six functions at 8017C0C8-8017C984 and move their shared types into a private header. This preserves all six exact matches while restoring the later gmregclear data anchors.

Replace the emitted data-order helper with named writable string arrays so the retail .data layout is retained without adding synthetic text.
