+++ src/melee/gm/gmregclear.c
@@ -1,5 +1,6 @@
 #include "gmregclear.h"
 
+#include "gmregclear_data.h"
 #include "platform.h"
 
 #include "baselib/forward.h"
@@ -146,46 +147,11 @@ struct lbl_80472E48_t {
 }; /* size = 0x80 */
 STATIC_ASSERT(sizeof(struct lbl_80472E48_t) == 0x80);
 
-/// Adventure mode stage data table entry (size 0x1A)
-/// Table has 110 entries: 22 stages × 5 difficulty levels
-/// Indexed as: difficulty * 5 + stage_slot
-typedef struct AdventureStageEntry {
-    /* 0x00 */ u8 stage_kind;
-    /* 0x01 */ u8 pad_1;
-    /* 0x02 */ u16 scale0_pct; ///< Divided by 100 to get ratio
-    /* 0x04 */ u16 scale1_pct; ///< Divided by 100 to get ratio
-    /* 0x06 */ u8 pad_6[0x14];
-} AdventureStageEntry;
-STATIC_ASSERT(sizeof(AdventureStageEntry) == 0x1A);
-
-/// Allstar mode stage data table entry (size 0x1A)
-/// Table has 55 entries: 11 stages × 5 difficulty levels
-typedef struct AllstarStageEntry {
-    /* 0x00 */ u8 stage_kind;
-    /* 0x01 */ u8 pad_1;
-    /* 0x02 */ u16 scale0_pct; ///< Divided by 100 to get ratio
-    /* 0x04 */ u16 scale1_pct; ///< Divided by 100 to get ratio
-    /* 0x06 */ u8 pad_6[0xA];
-    /* 0x10 */ u16 scale2_pct; ///< Divided by 100 to get ratio
-    /* 0x12 */ u16 scale3_pct; ///< Divided by 100 to get ratio
-    /* 0x14 */ u8 pad_14[0x6];
-} AllstarStageEntry;
-STATIC_ASSERT(sizeof(AllstarStageEntry) == 0x1A);
-
 typedef struct RegClearEv {
     /* 0x00 */ char pad_0[0x1C];
     /* 0x1C */ HSD_ImageDesc* x1C;
     /* 0x20 */ f32 x20;
 } RegClearEv;
-
-static struct lbl_804706C0_t {
-    int x0;
-    int x4;
-    int x8;
-    int xC; // HP (stamina)
-    int x10;
-    int x14;
-} lbl_804706C0;
 
 static UnkAdventureData lbl_80472C30;
 static UnkAllstarData lbl_80472CB0;
@@ -248,1092 +214,22 @@ STATIC_ASSERT(sizeof(lbl_80472ED8_t) == 0x6BC);
 
 lbl_80472ED8_t lbl_80472ED8;
 
-void fn_8017C0C8(void)
-{
-    PlayerInitData sp8;
-    gm_8016795C(&sp8);
-    sp8.slot_type = Gm_PKind_Cpu;
-    sp8.stocks = 1;
-    sp8.xE = 4;
-    sp8.cpu_level = Player_GetCpuLevel(2);
-    sp8.xE = Player_GetCpuType(2);
-    sp8.x18 = Player_GetAttackRatio(2);
-    sp8.x1C = Player_GetUnk50(2);
-    sp8.color = 0;
-    sp8.c_kind = CKIND_CREZYH;
-    sp8.team = 4;
-    sp8.xC_b7 = true;
-    sp8.hp = lbl_804706C0.xC;
-    sp8.xD_b2 = true;
-    sp8.xD_b0 = true;
-    sp8.xD_b2 = true;
-    sp8.spawn_dir = 1;
-    gm_8016EDDC(1, &sp8);
-}
-
-void fn_8017C1A4(HSD_GObj* unused)
-{
-    struct lbl_804706C0_t* tmp = &lbl_804706C0;
-    int temp_r29;
-    int temp_r29_2;
-    int temp_r28;
-    int temp_r27;
-    int temp_r27_2;
-    int temp_r31;
-    int temp_r3;
-    int temp_r3_3;
-    lbl_8046B6A0_t* temp_r3_2;
-    lbl_8046B6A0_t* temp_r3_4;
-    PAD_STACK(8);
-
-    temp_r31 = ftBossLib_8015C74C();
-    temp_r29 = ftBossLib_8015C7EC();
-    temp_r27 = ftBossLib_8015C88C();
-    temp_r3 = ftBossLib_8015C9CC();
-    switch (tmp->x0) {
-    case 0:
-        if (tmp->x8 == 0) {
-            temp_r3_2 = gm_16AE_GetUnkData_0();
-            temp_r3_2->x24C8.disable_pausing = false;
-            temp_r3_2->hud_enabled = 0;
-            lbAudioAx_800237A8(0x81650, 0x7F, 0x40);
-            ftLib_80086824();
-            tmp->x8 = 1;
-        } else if (ftBossLib_8015C270() == 0) {
-            tmp->x0 = 2;
-        }
-        break;
-    case 1:
-        if (tmp->x8 == 0) {
-            gm_16AE_GetUnkData_0();
-            lbAudioAx_800236DC();
-            ftBossLib_8015CA6C(0x1D9);
-            Player_SetMoreFlagsBit4(0, 1);
-            Player_SetMoreFlagsBit3(0, 1);
-            tmp->x10 = ft_80087C58(Player_GetEntity(0));
-            ft_80087C64(Player_GetEntity(0), 2);
-            tmp->x14 = ft_80087C58(Player_GetEntity(2));
-            ft_80087C64(Player_GetEntity(2), 2);
-            tmp->x8 = 1;
-        } else if (tmp->x8 == 1) {
-            lbAudioAx_800237A8(0x81651, 0x7F, 0x40);
-            fn_8017C0C8();
-            tmp->x8 = 2;
-        } else if (ftBossLib_8015C2A8() == 0) {
-            gm_16AE_GetUnkData_0();
-            Stage_80225074(0);
-            ftBossLib_8015CB7C();
-            Player_SetMoreFlagsBit4(0, 0);
-            Player_SetMoreFlagsBit3(0, 0);
-            ft_80087C64(Player_GetEntity(0), tmp->x10);
-            ft_80087C64(Player_GetEntity(2), tmp->x14);
-            ftBossLib_8015CC14();
-            tmp->x0 = 4U;
-        }
-        break;
-    case 2:
-        temp_r3_4 = gm_16AE_GetUnkData_0();
-        temp_r3_4->hud_enabled = 1;
-        Stage_80225074(0);
-        ftLib_800868A4();
-        ftBossLib_8015CC14();
-        tmp->x0 = 3;
-        break;
-    case 3:
-        if (Player_GetRemainingHP(2) <= 0) {
-            tmp->x0 = 7;
-            tmp->x8 = 0;
-        } else if (Player_GetRemainingHP(2) <= 0.5F * tmp->xC && fn_8017EDDC())
-        {
-            tmp->x0 = 1;
-            tmp->x8 = 0;
-        }
-        break;
-    case 4:
-        if (Player_GetRemainingHP(2) <= 0 && Player_GetRemainingHP(1) <= 0) {
-            tmp->x0 = 7;
-            tmp->x8 = 0;
-        } else if (Player_GetRemainingHP(2) <= 0 ||
-                   Player_GetRemainingHP(1) <= 0)
-        {
-            tmp->x0 = 6;
-            tmp->x8 = 0;
-        }
-        break;
-    case 5:
-        if (Player_GetRemainingHP(2) <= 0 && Player_GetRemainingHP(1) <= 0) {
-            tmp->x0 = 7;
-            tmp->x8 = 0;
-        }
-        break;
-    case 6:
-        if (Player_GetRemainingHP(2) <= 0) {
-            tmp->x4 = 2;
-        } else {
-            tmp->x4 = 1;
-        }
-        temp_r27_2 = temp_r31 + temp_r29;
-        if (tmp->x8 == 0) {
-            ftBossLib_8015CB9C(tmp->x4);
-            gm_801A4634(6);
-        } else if (tmp->x8 == temp_r31) {
-            lbAudioAx_8002438C(0x4E200);
-            ftBossLib_8015CC14();
-            gm_801A4674(6);
-            gm_SetGameSpeed(0.5f);
-        } else if (tmp->x8 >= temp_r27_2) {
-            gm_SetGameSpeed(1.0f);
-            tmp->x0 = 5;
-        }
-        if (Player_GetRemainingHP(2) <= 0 && Player_GetRemainingHP(1) <= 0) {
-            tmp->x0 = 7;
-            tmp->x8 = 0;
-        } else if (tmp->x8 <= temp_r27_2) {
-            tmp->x8++;
-        }
-        break;
-    case 7:
-        temp_r28 = temp_r31 + temp_r27;
-        temp_r29_2 = temp_r3 + temp_r28;
-        if (tmp->x8 == 0) {
-            temp_r3_4 = gm_16AE_GetUnkData_0();
-            temp_r3_4->hud_enabled = 0;
-            if (tmp->x4 == 2) {
-                tmp->x4 = 1;
-            } else {
-                tmp->x4 = 2;
-            }
-            ftBossLib_8015CB9C(tmp->x4);
-            gm_801A4634(6);
-            Player_80031790(0);
-            Player_80036844(0, 1);
-            temp_r3_4 = gm_16AE_GetUnkData_0();
-            temp_r3_4->x24C8.x0_6 = false;
-        } else if (tmp->x8 == temp_r31) {
-            lbAudioAx_8002438C(0x4E200);
-            ftBossLib_8015CC14();
-            gm_801A4674(6);
-            gm_SetGameSpeed(0.5f);
-            lbBgFlash_80020688(temp_r27);
-        } else if (tmp->x8 == temp_r28) {
-            lbBgFlash_800205F0(temp_r3);
-        } else if (tmp->x8 >= temp_r29_2) {
-            lbAudioAx_80028B90();
-            gm_SetGameSpeed(1.0f);
-            gm_8016B33C(8);
-            gm_8016B328();
-            break;
-        }
-        if (tmp->x8 >= temp_r31) {
-            lbAudioAx_80028B6C();
-        }
-        if (tmp->x8 <= temp_r28 && tmp->x8 % 30 == 0) {
-            gm_80167858(Player_GetPlayerId(0), Player_GetNametagSlotID(0), 2,
-                        0x1E);
-            Camera_80030E44(3, NULL);
-        }
-        if (tmp->x8 <= temp_r29_2) {
-            tmp->x8++;
-        }
-        break;
-    }
-}
-
-void fn_8017C71C(void)
-{
-    struct lbl_804706C0_t* tmp = &lbl_804706C0;
-    HSD_GObj_SetupProc(GObj_Create(0xF, 0x11, 0), fn_8017C1A4, 0x15);
-    tmp->x0 = 0;
-    tmp->x4 = 1;
-    tmp->x8 = 0;
-    tmp->xC = ftBossLib_8015C530(gm_8017E068());
-    Player_SetOtherStamina(2, tmp->xC);
-    ftLib_80087508(FTKIND_CREZYH, 0);
-}
-
-void fn_8017C7A0(void)
-{
-    lbBgFlash_8002063C(0x30);
-    if (Ground_801C1DAC()) {
-        gm_8016B350(0x7C85E);
-        return;
-    }
-    gm_8016B350(0x9C48);
-}
-
-void fn_8017C7EC(void)
-{
-    lbl_8046B6A0_t* temp_r31;
-    Unk1PData* temp_r30;
-
-    temp_r30 = fn_8017DF28();
-    temp_r31 = gm_16AE_GetUnkData_1();
-    temp_r31->timer_seconds = grPushOn_80219230(temp_r30->ckind);
-}
-
-void gm_8017C838(void)
-{
-    s8 sp10[3];
-
-    lbl_8046B6A0_t* temp_r3;
-    int i;
-    s8 var_r3;
-    s8* var_r31;
-    UnkAdventureData* temp_r30;
-
-    PAD_STACK(8);
-
-    temp_r30 = gm_GetAdventureData();
-    temp_r3 = gm_16AE_GetUnkData_0();
-    var_r31 = sp10;
-    sp10[0] = FTKIND_NONE;
-    sp10[1] = FTKIND_NONE;
-    sp10[2] = FTKIND_NONE;
-    switch (temp_r3->x24C8.xE) {
-    case 0x3B:
-        sp10[0] = FTKIND_YOSHI;
-        break;
-    case 0x3F:
-        sp10[0] = FTKIND_LINK;
-        break;
-    case 0x44:
-        sp10[0] = FTKIND_KIRBY;
-        if (temp_r30->x0.ckind == CKIND_KIRBY && temp_r30->x0.color == 0) {
-            var_r3 = 1;
-        } else {
-            var_r3 = 0;
-        }
-        ftLib_80087610(var_r3);
-        break;
-    case 0x48:
-        sp10[0] = 0xC;
-        if (gm_IsCKindUnlocked(CKIND_PICHU) != 0) {
-            sp10[1] = FTKIND_PICHU;
-        }
-        if (gm_IsCKindUnlocked(CKIND_PURIN) != 0) {
-            sp10[2] = FTKIND_PURIN;
-        }
-        break;
-    case 0x4C:
-        sp10[0] = FTKIND_POPO;
-        sp10[1] = FTKIND_NANA;
-        break;
-    case 0x4E:
-        sp10[0] = FTKIND_BOY;
-        sp10[1] = FTKIND_GIRL;
-        break;
-    }
-
-    for (i = 0; i < 3; i++) {
-        if (*var_r31 != FTKIND_MAX) {
-            ftLib_80087574(*var_r31);
-        }
-        var_r31++;
-    }
-    Camera_80030E34(0.2f);
-}
-
-/// Trailing pad keeps #lbl_803D85F0 at its retail offset.
-struct AdventureStageTable {
-    AdventureStageEntry entries[110];
-    u8 pad[4];
-} lbl_803D7AC0 = { {
-    { 0x04, 0x00, 40, 450, { 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 50, 400, { 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 60, 320, { 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 70, 240, { 0x05, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01, 0x00, 80, 170, { 0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 80, 140, { 0x00, 0x00, 0x10, 0x00, 0x00, 0x04, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 94, 112, { 0x03, 0x00, 0x10, 0x03, 0x00, 0x04, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 100, 107, { 0x05, 0x00, 0x10, 0x04, 0x00, 0x04, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 110, 105, { 0x07, 0x00, 0x10, 0x07, 0x00, 0x04, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 120, 95, { 0x08, 0x00, 0x10, 0x08, 0x00, 0x04, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 80, 140, { 0x01, 0x00, 0x04, 0x01, 0x01, 0x08, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 100, 90, { 0x02, 0x00, 0x04, 0x02, 0x01, 0x08, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 110, 70, { 0x04, 0x00, 0x04, 0x03, 0x01, 0x08, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 120, 50, { 0x07, 0x00, 0x04, 0x06, 0x01, 0x08, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01, 0x00, 130, 40, { 0x09, 0x00, 0x04, 0x09, 0x01, 0x08, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 60, 380, { 0x01, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x64, 0x00, 0x64,
-                             0x00, 0x00, 0x04, 0x00, 0x00, 0x04 } },
-    { 0x04, 0x00, 80, 250, { 0x03, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x64, 0x00, 0x64,
-                             0x00, 0x00, 0x04, 0x00, 0x00, 0x04 } },
-    { 0x03, 0x00, 90, 200, { 0x05, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x64, 0x00, 0x64,
-                             0x00, 0x00, 0x04, 0x00, 0x00, 0x04 } },
-    { 0x02, 0x00, 95, 160, { 0x07, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x64, 0x00, 0x64,
-                             0x00, 0x00, 0x04, 0x00, 0x00, 0x04 } },
-    { 0x01, 0x00, 110, 120, { 0x08, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x64, 0x00, 0x64,
-                              0x00, 0x00, 0x04, 0x00, 0x00, 0x04 } },
-    { 0x03, 0x00, 70, 160, { 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 80, 140, { 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 90, 120, { 0x05, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01, 0x00, 100, 105, { 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01, 0x00, 108, 95, { 0x09, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 80, 125, { 0x02, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 90, 110, { 0x06, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 100, 100, { 0x08, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 118, 95, { 0x09, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 130, 90, { 0x09, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 80, 130, { 0x02, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 90, 110, { 0x05, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 100, 100, { 0x07, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 115, 88, { 0x08, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01, 0x00, 130, 74, { 0x09, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 0, 0, { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 0, 0, { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 0, 0, { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 0, 0, { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 0, 0, { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 80, 170, { 0x02, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 100, 130, { 0x04, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 120, 110, { 0x06, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01, 0x00, 130, 100, { 0x08, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 140, 90, { 0x09, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 60, 500, { 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 70, 400, { 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 80, 300, { 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01, 0x00, 90, 200, { 0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 100, 175, { 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 70, 240, { 0x02, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 75, 180, { 0x04, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 80, 130, { 0x06, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01, 0x00, 100, 100, { 0x07, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01, 0x00, 120, 90, { 0x09, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 80, 180, { 0x02, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 90, 150, { 0x04, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 100, 130, { 0x05, 0x00, 0x12, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 118, 110, { 0x06, 0x00, 0x12, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 130, 100, { 0x07, 0x00, 0x12, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 60, 210, { 0x02, 0x01, 0x08, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 70, 180, { 0x04, 0x01, 0x08, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 90, 140, { 0x06, 0x01, 0x08, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 105, 120, { 0x08, 0x01, 0x08, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 117, 100, { 0x09, 0x01, 0x08, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 60, 420, { 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 78, 320, { 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 88, 200, { 0x05, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 98, 170, { 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 109, 150, { 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 0, 0, { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 0, 0, { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 0, 0, { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 0, 0, { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 0, 0, { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 80, 150, { 0x03, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 100, 110, { 0x06, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 110, 100, { 0x08, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 120, 90, { 0x09, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 130, 70, { 0x09, 0x02, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 70, 190, { 0x01, 0x00, 0x1D, 0x01, 0x02, 0x1D, 0x01,
-                             0x01, 0x1B, 0x00, 0x00, 0x64, 0x00, 0x64,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 80, 150, { 0x03, 0x00, 0x1D, 0x03, 0x02, 0x1D, 0x02,
-                             0x01, 0x1D, 0x00, 0x00, 0x64, 0x00, 0x64,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 90, 130, { 0x05, 0x00, 0x1D, 0x04, 0x02, 0x1D, 0x04,
-                             0x01, 0x1D, 0x00, 0x00, 0x64, 0x00, 0x64,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 100, 110, { 0x07, 0x00, 0x1D, 0x06, 0x02, 0x1D, 0x05,
-                              0x01, 0x1D, 0x00, 0x00, 0x64, 0x00, 0x64,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04, 0x00, 130, 100, { 0x08, 0x00, 0x1D, 0x08, 0x02, 0x1D, 0x08,
-                              0x01, 0x1D, 0x00, 0x00, 0x64, 0x00, 0x64,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 80, 190, { 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 85, 140, { 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 90, 130, { 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 95, 120, { 0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 100, 100, { 0x09, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 50, 420, { 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 60, 370, { 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 70, 280, { 0x05, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 80, 200, { 0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 90, 150, { 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 70, 190, { 0x01, 0x00, 0x1B, 0x01, 0x00, 0x1B, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 90, 150, { 0x03, 0x00, 0x1B, 0x02, 0x00, 0x1B, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 100, 130, { 0x05, 0x00, 0x1B, 0x03, 0x00, 0x1B, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 110, 110, { 0x06, 0x00, 0x1B, 0x05, 0x00, 0x1B, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 120, 90, { 0x08, 0x00, 0x1B, 0x07, 0x00, 0x1B, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03, 0x00, 70, 130, { 0x02, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 93, 115, { 0x04, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 100, 105, { 0x06, 0x03, 0x04, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                              0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01, 0x00, 110, 88, { 0x08, 0x03, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 120, 75, { 0x09, 0x03, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 1, 1, { 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02, 0x00, 1, 1, { 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                          0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01, 0x00, 135, 58, { 0x07, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01, 0x00, 148, 50, { 0x08, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF, 0x00, 160, 38, { 0x09, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-} };
-
-/// Trailing pad keeps #lbl_803D8B88 at its retail offset.
-struct AllstarStageTable {
-    AllstarStageEntry entries[55];
-    u8 pad[2];
-} lbl_803D85F0 = { {
-    { 0x04,
-      0x00,
-      60,
-      175,
-      { 0x00, 0x00, 0x09, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03,
-      0x00,
-      85,
-      120,
-      { 0x02, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03,
-      0x00,
-      100,
-      100,
-      { 0x04, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      100,
-      100,
-      { 0x07, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01,
-      0x00,
-      120,
-      90,
-      { 0x09, 0x03, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04,
-      0x00,
-      70,
-      160,
-      { 0x01, 0x00, 0x04, 0x02, 0x00, 0x09, 0x00, 0x00, 0x00, 0x00 },
-      80,
-      90,
-      { 0x08, 0x01, 0x04, 0x00, 0x00, 0x00 } },
-    { 0x04,
-      0x00,
-      85,
-      120,
-      { 0x03, 0x00, 0x04, 0x03, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00 },
-      80,
-      100,
-      { 0x07, 0x01, 0x04, 0x00, 0x00, 0x00 } },
-    { 0x03,
-      0x00,
-      100,
-      100,
-      { 0x05, 0x00, 0x04, 0x04, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00 },
-      80,
-      115,
-      { 0x06, 0x01, 0x04, 0x00, 0x00, 0x00 } },
-    { 0x03,
-      0x00,
-      100,
-      90,
-      { 0x06, 0x00, 0x04, 0x06, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00 },
-      80,
-      130,
-      { 0x04, 0x01, 0x04, 0x00, 0x00, 0x00 } },
-    { 0x01,
-      0x00,
-      115,
-      80,
-      { 0x08, 0x03, 0x04, 0x08, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00 },
-      80,
-      150,
-      { 0x03, 0x01, 0x04, 0x00, 0x00, 0x00 } },
-    { 0xFF,
-      0x00,
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF,
-      0x00,
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF,
-      0x00,
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF,
-      0x00,
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF,
-      0x00,
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04,
-      0x00,
-      75,
-      140,
-      { 0x02, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04,
-      0x00,
-      90,
-      110,
-      { 0x05, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04,
-      0x00,
-      100,
-      100,
-      { 0x08, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04,
-      0x00,
-      115,
-      85,
-      { 0x09, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x04,
-      0x00,
-      134,
-      70,
-      { 0x09, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03,
-      0x00,
-      70,
-      360,
-      { 0x02, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      100,
-      100,
-      { 0x06, 0x02, 0x04, 0x06, 0x01, 0x04 } },
-    { 0x03,
-      0x00,
-      80,
-      240,
-      { 0x03, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      100,
-      90,
-      { 0x05, 0x02, 0x04, 0x05, 0x01, 0x04 } },
-    { 0x02,
-      0x00,
-      90,
-      180,
-      { 0x05, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      100,
-      80,
-      { 0x02, 0x02, 0x04, 0x02, 0x01, 0x09 } },
-    { 0x02,
-      0x00,
-      100,
-      140,
-      { 0x07, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      100,
-      80,
-      { 0x01, 0x02, 0x04, 0x01, 0x01, 0x09 } },
-    { 0x01,
-      0x00,
-      120,
-      100,
-      { 0x08, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      100,
-      80,
-      { 0x00, 0x02, 0x04, 0x00, 0x01, 0x09 } },
-    { 0xFF,
-      0x00,
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF,
-      0x00,
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF,
-      0x00,
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF,
-      0x00,
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF,
-      0x00,
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      85,
-      130,
-      { 0x03, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      100,
-      110,
-      { 0x06, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      90,
-      130,
-      { 0x05, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      100,
-      110,
-      { 0x06, 0x00, 0x12, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01,
-      0x00,
-      110,
-      90,
-      { 0x08, 0x00, 0x12, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03,
-      0x00,
-      60,
-      400,
-      { 0x01, 0x00, 0x19, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x03,
-      0x00,
-      70,
-      260,
-      { 0x03, 0x00, 0x17, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      80,
-      200,
-      { 0x04, 0x00, 0x17, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      90,
-      155,
-      { 0x05, 0x00, 0x17, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01,
-      0x00,
-      110,
-      120,
-      { 0x07, 0x00, 0x17, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      100,
-      100,
-      { 0x01, 0x00, 0x04, 0x00, 0x00, 0x04, 0x00, 0x00, 0x04, 0x00 },
-      100,
-      100,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      100,
-      100,
-      { 0x01, 0x00, 0x04, 0x00, 0x00, 0x04, 0x00, 0x00, 0x04, 0x00 },
-      100,
-      100,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      100,
-      100,
-      { 0x01, 0x00, 0x04, 0x00, 0x00, 0x04, 0x00, 0x00, 0x04, 0x00 },
-      100,
-      100,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      100,
-      100,
-      { 0x01, 0x00, 0x04, 0x00, 0x00, 0x04, 0x00, 0x00, 0x04, 0x00 },
-      100,
-      100,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      100,
-      100,
-      { 0x01, 0x00, 0x04, 0x00, 0x00, 0x04, 0x00, 0x00, 0x04, 0x00 },
-      100,
-      100,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      80,
-      180,
-      { 0x01, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      90,
-      130,
-      { 0x03, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      100,
-      100,
-      { 0x05, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x01,
-      0x00,
-      110,
-      85,
-      { 0x08, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF,
-      0x00,
-      125,
-      68,
-      { 0x09, 0x00, 0x1B, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      50,
-      100,
-      { 0x05, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      75,
-      100,
-      { 0x06, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      100,
-      100,
-      { 0x07, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0x02,
-      0x00,
-      150,
-      100,
-      { 0x08, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-    { 0xFF,
-      0x00,
-      200,
-      100,
-      { 0x09, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 },
-      0,
-      0,
-      { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 } },
-} };
-
 u16 lbl_803D8B88[] = { 0x18, 0x16, 0x12, 0x3, 0x5, 0x4, 0x6, 0x1a, 0x19, 0x7 };
 
-/// @todo .data order hack
-static void order_data(void)
-{
-    (void) "GmRegClr";
-    (void) "ScGamRegClear_scene_data";
-    (void) "Error : Cannot open archive file (File Name : %s).";
-    (void) "SdDec.usd";
-    (void) "SIS_DecisionData";
-    (void) "SdDec.dat";
-    (void) "          ";
-    (void) "               ";
-    (void) "Error : gobj don\'t get (gmRegClearAddModel)\n";
-    (void) "gmregclear.c";
-    (void) "Error : jobj don\'t get (gmRegClearAddModel)\n";
-}
+static char gmRegClearArchiveName[] = "GmRegClr";
+static char gmRegClearSceneData[] = "ScGamRegClear_scene_data";
+static char gmRegClearArchiveError[] =
+    "Error : Cannot open archive file (File Name : %s).";
+static char gmRegClearDecisionUsd[] = "SdDec.usd";
+static char gmRegClearDecisionData[] = "SIS_DecisionData";
+static char gmRegClearDecisionDat[] = "SdDec.dat";
+static char gmRegClearShortBlank[] = "          ";
+static char gmRegClearLongBlank[] = "               ";
+static char gmRegClearGObjError[] =
+    "Error : gobj don\'t get (gmRegClearAddModel)\n";
+static char gmRegClearFilename[] = "gmregclear.c";
+static char gmRegClearJObjError[] =
+    "Error : jobj don\'t get (gmRegClearAddModel)\n";
 
 void gm_8017C984(UNK_T arg0)
 {
@@ -2754,9 +1650,9 @@ s32 fn_8017F2A4(HSD_Text** arg0, f32 farg0, f32 farg1)
 
     temp = HSD_SisLib_803A611C(3, NULL, 9U, 0xDU, 0U, 0x14U, 0U, 0x13U);
     if (lbLang_IsSavedLanguageUS()) {
-        HSD_SisLib_803A62A0(3, "SdDec.usd", "SIS_DecisionData");
+        HSD_SisLib_803A62A0(3, gmRegClearDecisionUsd, gmRegClearDecisionData);
     } else {
-        HSD_SisLib_803A62A0(3, "SdDec.dat", "SIS_DecisionData");
+        HSD_SisLib_803A62A0(3, gmRegClearDecisionDat, gmRegClearDecisionData);
     }
     *arg0 = HSD_SisLib_803A6754(3, temp);
     ptr = arg0;
@@ -2774,11 +1670,11 @@ s32 fn_8017F2A4(HSD_Text** arg0, f32 farg0, f32 farg1)
         text->x34.x = 0.56f;
         text->x34.y = 0.6f;
         ptr[1]->default_kerning = 1;
-        HSD_SisLib_803A6B98(*arg0, x_end, -12.0f + y, "          ");
+        HSD_SisLib_803A6B98(*arg0, x_end, -12.0f + y, gmRegClearShortBlank);
         ptr++;
     }
     return HSD_SisLib_803A6B98(*arg0, x_end, 9.0f + (134.40001f + farg1),
-                               "               ");
+                               gmRegClearLongBlank);
     PAD_STACK(8);
 }
 
@@ -3316,8 +2212,6 @@ void fn_8017FF1C(HSD_GObj* gobj)
     PAD_STACK(0x1C);
 }
 
-/// @todo Retail anchors this data pool at #lbl_803D8B88, suggesting the
-///       original file was split into separate translation units here.
 s32 fn_801803FC(void* arg0)
 {
     fn_8017FA1C_arg* p = arg0;
@@ -3334,13 +2228,13 @@ s32 fn_801803FC(void* arg0)
     p->x0 = gobj;
     if (gobj == NULL) {
         HSD_JObjAnimAll((HSD_JObj*) gobj->hsd_obj);
-        OSReport("Error : gobj don\'t get (gmRegClearAddModel)\n");
-        OSPanic(__FILE__, 0x42C, "");
+        OSReport(gmRegClearGObjError);
+        OSPanic(gmRegClearFilename, 0x42C, "");
     }
     jobj = HSD_JObjLoadJoint(mdl->joint);
     if (jobj == NULL) {
-        OSReport("Error : jobj don\'t get (gmRegClearAddModel)\n");
-        OSPanic(__FILE__, 0x432, "");
+        OSReport(gmRegClearJObjError);
+        OSPanic(gmRegClearFilename, 0x432, "");
     }
     HSD_GObjObject_80390A70(gobj, HSD_GObj_804D7849, jobj);
     GObj_SetupGXLink(gobj, HSD_GObj_JObjCallback, 0xBU, 0U);
@@ -3389,8 +2283,7 @@ static inline void fn_80180630_CheckArchive(HSD_Archive* archive, void* sp38)
     struct lbl_80472D28_t* state = &lbl_80472D28;
     state->x48 = archive;
     if (sp38 == NULL) {
-        OSReport("Error : Cannot open archive file (File Name : %s).",
-                 "GmRegClr");
+        OSReport(gmRegClearArchiveError, gmRegClearArchiveName);
     }
 }
 
@@ -3524,12 +2417,11 @@ void fn_80180630(int arg0, int arg1, int arg2, bool arg3,
     state->x104 = var_r4;
     lbl_804D65C0 = (var_r4 - (arg0 + arg1)) / 10;
 
-    archive =
-        lbArchive_80016DBC("GmRegClr", &sp38, "ScGamRegClear_scene_data", 0);
+    archive = lbArchive_80016DBC(gmRegClearArchiveName, &sp38,
+                                 gmRegClearSceneData, 0);
     state->x48 = archive;
     if (sp38 == NULL) {
-        OSReport("Error : Cannot open archive file (File Name : %s).",
-                 "GmRegClr");
+        OSReport(gmRegClearArchiveError, gmRegClearArchiveName);
     }
     fn_80168A6C(sp38, &state->x4C, 0);
